Add unit tests and rework almost everything
Summary:
Achieve 100% unit test coverage because it would be hypocritical to
write a testing tool that didn't test itself. `setup.py test` or `tox`
will test against python 2 and 3 in a virtualenv. This required a
refactor of basically everything to add explicit interfaces and allow
dependencies to be injected.

The CLI interface is consolidated into just a `cinderblock` command,
with subcommands `cinderblock trigger` and `cinderblock update`. This
should make it easier to extend the program without a proliferation of
commands in the future. Resolves T559.

Test Plan: https://circleci.com/gh/bitglue/cinderblock/6

Reviewers: #developers, dong

Reviewed By: #developers, dong

Subscribers: dong

Maniphest Tasks: T559

Differential Revision: https://phabricator.paperg.com/D349