Don't pass CircleCI API token to triggered build
Summary:
This mostly reverts 465199fb7dc9b61c6151bbcc89b8db4d670e43e7.

We decided this wasn't a great idea, because it requires each project to
duplicate the same key, only to pass it to the integration project. It's
easier to just put the key in the integration project and not pass it
anywhere. Less duplication.

See also D362.

Test Plan: python setup.py test

Reviewers: #developers, dong

Reviewed By: #developers, dong

Subscribers: dong

Differential Revision: https://phabricator.paperg.com/D363