Print URL for triggered build
Summary:
Where we used to print a link to builds for the triggered branch
generally, now print a URL to the specific build that was triggered.
This will make it easier when looking at a source build to find the
triggered build.

Test Plan:
Updated unit tests. Also ran a
`cinderblock trigger paperg/integration/fork` locally, and verified that
the printed URL was correct.

Reviewers: #developers, dong

Reviewed By: #developers, dong

Differential Revision: https://phabricator.paperg.com/D421